improvement: add validation zod
