+++ apps/web/backoffice/package.json
@@ -24,6 +24,7 @@
     "superjson": "2.2.1",
     "ts-pattern": "5.0.8",
     "tslib": "2.6.2",
-    "typescript": "5.3.3"
+    "typescript": "5.3.3",
+    "zod": "3.22.4"
   }
 }
