--- libs/api/src/activity/modules/activity.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ActivityService } from '../services';
-import { ActivityController } from '../controllers';
-
-@Module({
-  imports: [],
-  controllers: [ActivityController],
-  providers: [ActivityService],
-})
-export class ActivityModule {}
