+++ libs/api/src/activity/modules/index.ts
@@ -1,1 +1,10 @@
-export * from './activity.module';
+import { Module } from '@nestjs/common';
+import { ActivityService } from '../services';
+import { ActivityController } from '../controllers';
+
+@Module({
+  imports: [],
+  controllers: [ActivityController],
+  providers: [ActivityService],
+})
+export class ActivityModule {}
