--- libs/api/src/activity/routers/activity.router.ts
@@ -1,4 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class ActivityRouter {}
