+++ libs/api/src/user/modules/index.ts
@@ -1,1 +1,11 @@
-export * from './user.module';
+import { Module } from '@nestjs/common';
+
+import { UserService } from '../services';
+import { UserController } from '../controllers';
+
+@Module({
+  imports: [],
+  controllers: [UserController],
+  providers: [UserService],
+})
+export class UserModule {}
