--- libs/api/src/user/modules/user.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { UserService } from '../services';
-import { UserController } from '../controllers';
-
-@Module({
-  imports: [],
-  controllers: [UserController],
-  providers: [UserService],
-})
-export class UserModule {}
