+++ libs/entities/src/index.ts
@@ -1,2 +1,3 @@
 export * from './types';
 export * from './enums';
+export * from './schemas';
