+++ libs/entities/src/schemas/index.ts
@@ -1,0 +1,3 @@
+export * from "./activity"
+export * from "./user"
+export * from "./auth"
