+++ package.json
@@ -68,6 +68,7 @@
     "clsx": "^2.1.0",
     "drizzle-kit": "^0.20.14",
     "drizzle-orm": "^0.29.3",
+    "express": "^4.18.2",
     "next": "14.0.4",
     "next-auth": "^4.24.6",
     "passport-google-oauth20": "^2.0.0",
